Compiling on Windows.

+++ slot_machine/src/roll/NextStepRoll.cpp
@@ -3,7 +3,7 @@
 NextStepRoll::NextStepRoll(int nframes, const std::vector<int> &value_counts)
         : ExactRoll(static_cast<int>(value_counts.size()), nframes)
 {
-    for (int i = 0; i < value_counts.size(); ++i) {
+    for (size_t i = 0; i < value_counts.size(); ++i) {
         set_speed(i, 1. / (value_counts[i] * nframes));
     }
 }
